data/requirements.txt
pandas==2.2.0
requests==2.31.0
PyYAML==6.0.1
SQLAlchemy==2.0.25
pandas-market-calendars==4.3.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
ccxt==4.1.13